+++ Coosu.Storyboard/Common/ICommonEvent.cs
@@ -3,8 +3,8 @@
     public interface ICommonEvent : IEvent, IScriptable
     {
         EasingType Easing { get; set; }
-        float[] Start { get; set; }
-        float[] End { get; set; }
+        double[] Start { get; set; }
+        double[] End { get; set; }
         int ParamLength { get; }
         bool IsStatic { get; }
     }
