+++ Coosu.Storyboard/Common/ISceneObject.cs
@@ -5,9 +5,9 @@ namespace Coosu.Storyboard.Common
 {
     public interface ISceneObject : IDefinedObject, IEventHost
     {
-        public float DefaultY { get; set; }
-        public float DefaultX { get; set; }
-        public float ZDistance { get; }
+        public double DefaultY { get; set; }
+        public double DefaultX { get; set; }
+        public double ZDistance { get; }
         public int CameraId { get; }
         List<Loop> LoopList { get; }
         List<Trigger> TriggerList { get; }
