+++ Coosu.Storyboard/Events/Parameter.cs
@@ -26,8 +26,8 @@ namespace Coosu.Storyboard.Events
             await writer.WriteAsync(Type.ToShortString());
         }
 
-        public Parameter(EasingType easing, float startTime, float endTime, ParameterType type) :
-            base(easing, startTime, endTime, new float[] { (int)type }, new float[] { (int)type })
+        public Parameter(EasingType easing, double startTime, double endTime, ParameterType type) :
+            base(easing, startTime, endTime, new double[] { (int)type }, new double[] { (int)type })
         {
             Easing = EasingType.Linear;
         }
